refactor: AI 패널 채팅 훅 책임 분리
